+++ CCID/ccid_lib/ccid_openpgp_key.h
@@ -27,10 +27,10 @@ typedef enum {
 //
 // 関数群
 //
-uint16_t    openpgp_key_get_attributes(uint8_t tag, uint8_t *buf, size_t *size);
-uint16_t    openpgp_key_get_fingerprint(uint8_t tag, void *buf, size_t *size);
-uint16_t    openpgp_key_get_datetime(uint8_t tag, void *buf, size_t *size);
-uint16_t    openpgp_key_get_status(uint8_t tag, uint8_t *status);
+uint16_t    openpgp_key_get_attributes(uint16_t tag, uint8_t *buf, size_t *size);
+uint16_t    openpgp_key_get_fingerprint(uint16_t tag, void *buf, size_t *size);
+uint16_t    openpgp_key_get_datetime(uint16_t tag, void *buf, size_t *size);
+uint16_t    openpgp_key_get_status(uint16_t tag, uint8_t *status);
 
 #ifdef __cplusplus
 }
